// front/lib/generated/transfer_hook/types/restrictionMode.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum RestrictionMode {
  Open,
  KycGated,
}

export type RestrictionModeArgs = RestrictionMode;

export function getRestrictionModeEncoder(): FixedSizeEncoder<RestrictionModeArgs> {
  return getEnumEncoder(RestrictionMode);
}

export function getRestrictionModeDecoder(): FixedSizeDecoder<RestrictionMode> {
  return getEnumDecoder(RestrictionMode);
}

export function getRestrictionModeCodec(): FixedSizeCodec<
  RestrictionModeArgs,
  RestrictionMode
> {
  return combineCodec(getRestrictionModeEncoder(), getRestrictionModeDecoder());
}